rename jsx to js files
