--- docs/index.jsx
@@ -1,8 +1,0 @@
-'use strict';
-
-import React from 'react'
-import ReactDOM from 'react-dom'
-import Docs from './docs'
-
-React.initializeTouchEvents && React.initializeTouchEvents(true);
-ReactDOM.render(<Docs />, document.getElementById('rapp'));
